Refactor due to circular import
