+++ great_expectations/core/expectation_validation_result.py
@@ -13,10 +13,12 @@ from great_expectations.core.expectation_configuration import (
 from great_expectations.core.util import (
     convert_to_json_serializable,
     ensure_json_serializable,
-    get_atomic_rendered_content_for_object,
     in_jupyter_notebook,
 )
 from great_expectations.marshmallow__shade import Schema, fields, post_load, pre_dump
+from great_expectations.render.renderer.inline_renderer import (
+    get_atomic_rendered_content_for_object,
+)
 from great_expectations.render.types import RenderedAtomicContentSchema, RenderedContent
 from great_expectations.types import SerializableDictDot
 
@@ -198,7 +200,7 @@ class ExpectationValidationResult(SerializableDictDot):
         self.rendered_content: List[
             RenderedContent
         ] = get_atomic_rendered_content_for_object(
-            object=self,
+            render_object=self,
         )
 
     @staticmethod
