+++ great_expectations/core/util.py
@@ -8,7 +8,7 @@ import sys
 import uuid
 from collections import OrderedDict
 from collections.abc import Mapping
-from typing import Any, Callable, Dict, Iterable, List, Optional, Union
+from typing import Any, Dict, Iterable, List, Optional, Union
 from urllib.parse import urlparse
 
 import dateutil.parser
@@ -17,17 +17,8 @@ import pandas as pd
 from IPython import get_ipython
 
 from great_expectations import exceptions as ge_exceptions
-from great_expectations.core.expectation_configuration import ExpectationConfiguration
-from great_expectations.core.expectation_validation_result import (
-    ExpectationValidationResult,
-)
 from great_expectations.core.run_identifier import RunIdentifier
 from great_expectations.exceptions import InvalidExpectationConfigurationError
-from great_expectations.expectations.registry import (
-    get_renderer_impl,
-    get_renderer_names_with_renderer_prefix,
-)
-from great_expectations.render.types import RenderedContent
 from great_expectations.types import SerializableDictDot
 from great_expectations.types.base import SerializableDotDict
 
@@ -785,68 +776,3 @@ def get_sql_dialect_floating_point_infinity_value(
             return res["NegativeInfinity"]
         else:
             return res["PositiveInfinity"]
-
-
-def get_atomic_rendered_content_for_object(
-    object: Union[ExpectationConfiguration, ExpectationValidationResult],
-) -> List[RenderedContent]:
-    """Gets RenderedAtomicContent for a given ExpectationConfiguration or ExpectationValidationResult.
-
-    Args:
-        object: An ExpectationConfiguration or ExpectationValidationResult for which to return rendered_content.
-
-    Returns:
-        A list of RenderedAtomicContent objects for a given ExpectationConfiguration or ExpectationValidationResult.
-    """
-    expectation_type: str
-    atomic_renderer_prefix: str
-    legacy_renderer_prefix: str
-    if isinstance(object, ExpectationConfiguration):
-        expectation_type = object.expectation_type
-        atomic_renderer_prefix = "atomic.prescriptive"
-        legacy_renderer_prefix = "renderer.prescriptive"
-    elif isinstance(object, ExpectationValidationResult):
-        expectation_type = object.expectation_config.expectation_type
-        atomic_renderer_prefix = "atomic.diagnostic"
-        legacy_renderer_prefix = "renderer.diagnostic"
-    else:
-        raise ValueError(
-            f"object must be of type ExpectationConfiguration or ExpectationValidationResult, but an object of type {type(object)} was passed"
-        )
-
-    renderer_names: List[str] = get_renderer_names_with_renderer_prefix(
-        object_name=expectation_type,
-        renderer_prefix=atomic_renderer_prefix,
-    )
-    if len(renderer_names) == 0:
-        renderer_names = get_renderer_names_with_renderer_prefix(
-            object_name=expectation_type,
-            renderer_prefix=legacy_renderer_prefix,
-        )
-
-    renderer_tuple: Optional[tuple]
-    renderer_fn: Callable
-    renderer_rendered_content: Optional[
-        Union[RenderedContent, List[RenderedContent]]
-    ] = None
-    rendered_content: List[RenderedContent] = []
-    for renderer_name in renderer_names:
-        renderer_tuple = get_renderer_impl(
-            object_name=expectation_type, renderer_type=renderer_name
-        )
-        if renderer_tuple is not None:
-            # index 0 is expectation class-name and index 1 is implementation of renderer
-            renderer_fn = renderer_tuple[1]
-            if isinstance(object, ExpectationConfiguration):
-                renderer_rendered_content = renderer_fn(configuration=object)
-            elif isinstance(object, ExpectationValidationResult):
-                renderer_rendered_content = renderer_fn(result=object)
-
-            if renderer_rendered_content is None:
-                pass
-            elif isinstance(renderer_rendered_content, list):
-                rendered_content.extend(renderer_rendered_content)
-            else:
-                rendered_content.append(renderer_rendered_content)
-
-    return rendered_content
